LPD-27959 Replace the editingLanguageId only at the end of the name
Before we were replacing the editingLanguageId everytime there was a match on the string, which is not a problem to our languages out of the box that follow a predefined and distinctive format. But when dealing with custom languages, which are only formed by two lowercase letters with no symbols, the function was replacing other parts of the name that weren't language ids at all. This change is to make sure we are only replacing the previous language id for the new one located at the end of the string, without compromising other parts of the name.